Update MCP integration
