--- mcp/livekit-mcp-server/pyproject.toml
@@ -1,13 +1,0 @@
-[project]
-name = "livekit-mcp"
-version = "0.1.0"
-description = "LiveKit MCP tools for managing rooms, tokens, and participant interactions"
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "livekit>=0.21.3",
-    "livekit-api>=0.8.2",
-    "mcp[cli]>=1.4.1",
-    "nest-asyncio>=1.6.0",
-    "python-dotenv>=1.0.0",
-]
